+++ srcore/srio/sr-io-ext.go
@@ -12,40 +12,40 @@ import "github.com/Joystream/tinygo-wasm-substrate/srcore/primitives"
 // TODO: all these should be unexported, export only high-level wrappers
 
 //go:export ext_clear_prefix
-func ext_clear_prefix(prefix_data *byte, prefix_len uint32)
+func ext_clear_prefix(prefix_data *byte, prefix_len uintptr)
 
 //go:export ext_print_utf8
-func ext_print_utf8(utf8_data *byte, utf8_len uint32)
+func ext_print_utf8(utf8_data *byte, utf8_len uintptr)
 
 //go:export ext_set_storage
-func ext_set_storage(key_data *byte, key_len uint32, value_data *byte, value_len uint32)
+func ext_set_storage(key_data *byte, key_len uintptr, value_data *byte, value_len uintptr)
 
 //go:export ext_get_allocated_storage
-func ext_get_allocated_storage(key_data *byte, key_len uint32, value_len_ptr *uint32) *byte
+func ext_get_allocated_storage(key_data *byte, key_len uintptr, value_len_ptr *uintptr) *byte
 
 //go:export ext_clear_storage
-func ext_clear_storage(key_data *byte, key_len uint32)
+func ext_clear_storage(key_data *byte, key_len uintptr)
 
 //go:export ext_blake2_256
-func ext_blake2_256(data *byte, len uint32, out *byte)
+func ext_blake2_256(data *byte, len uintptr, out *byte)
 
 //go:export ext_twox_128
-func ext_twox_128(data *byte, len uint32, out *byte)
+func ext_twox_128(data *byte, len uintptr, out *byte)
 
 //go:export ext_twox_256
-func ext_twox_256(data *byte, len uint32, out *byte)
+func ext_twox_256(data *byte, len uintptr, out *byte)
 
 //go:export ext_ed25519_verify
-func Ext_ed25519_verify(msg_data *byte, msg_len uint32, sig_data *byte, pubkey_data *byte) uint32
+func Ext_ed25519_verify(msg_data *byte, msg_len uintptr, sig_data *byte, pubkey_data *byte) uint32
 
 //go:export ext_blake2_256_enumerated_trie_root
-func ext_blake2_256_enumerated_trie_root(values_data *byte, lens_data_addr *uint32, lens_len uint32, resultPtr *byte)
+func ext_blake2_256_enumerated_trie_root(values_data *byte, lens_data_addr *uintptr, lens_len uintptr, resultPtr *byte)
 
 //go:export ext_storage_root
 func ext_storage_root(resultPtr *primitives.H256)
 
 //go:export ext_storage_changes_root
-func ext_storage_changes_root(parent_hash_data *byte, parent_hash_len uint32, parent_num uint64, result *primitives.H256) uint32
+func ext_storage_changes_root(parent_hash_data *byte, parent_hash_len uintptr, parent_num uint64, result *primitives.H256) uint32
 
 //go:export ext_clear_prefix
-func Ext_clear_prefix(prefix_data *byte, prefix_len uint32)
+func Ext_clear_prefix(prefix_data *byte, prefix_len uintptr)
